refactor: moved inAppUpdate class to utils package
